+++ src/dev/drivers/mod.rs
@@ -2,4 +2,6 @@
 //
 // This is the central mod.rs file for the drivers subdirectory of the dev directory.
 
+//pub mod mm_uart;
 pub mod pic8259;
+pub mod uart;