+++ src/math/float/mod.rs
@@ -3,6 +3,8 @@
 // This is the mod.rs file for the math::float module.
 
 pub mod absval;
+pub mod cosine;
+pub mod cpsign;
 pub mod fl;
 pub mod floor;
 pub mod hypotenuse;
@@ -10,5 +12,10 @@ pub mod icos;
 pub mod invert;
 pub mod itan;
 pub mod log;
+pub mod logb2;
+pub mod logb10;
 pub mod nlog;
+pub mod round;
+pub mod sine;
 pub mod sr;
+pub mod tangent;